feat: remove exception add runtimeexception

+++ src/main/java/com/nayoon/ai_shop/controller/request/StockRequest.java
@@ -2,13 +2,13 @@ package com.nayoon.ai_shop.controller.request;
 
 public class StockRequest {
     private Long productId;
-    private int quantity;
+    private Long quantity;
 
     public Long getProductId() {
         return productId;
     }
 
-    public int getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 }
